requirements.txt
beautifulsoup4==4.12.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.1.0
idna==3.4
IMAPClient==2.3.1
pyocclient==0.6
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
